admin/sl: Add err library
This new library is used to extract results from the Shadow Link service
and throw the appropriate exception upon error.
